Move mobile session header button out of routes
