+++ mobile/src/session/MobileSessionHeaderIconButton.tsx
@@ -1,7 +1,7 @@
-import { Pressable } from 'react-native'
 import type { ComponentType } from 'react'
-import { colors } from '../../../../src/theme/mobile-theme'
-import { styles } from './mobile-session-styles'
+import { Pressable } from 'react-native'
+import { colors } from '../theme/mobile-theme'
+import { styles } from '../../app/h/[hostId]/session/mobile-session-styles'
 
 type HeaderIconProps = {
   size?: number
